from tensorflow docs
The expected return values are: features: A dict of string to Tensor or SparseTensor, specifying the features to be passed to the model. labels: A Tensor, SparseTensor, or a dict of string to Tensor or SparseTensor, specifying labels for training or eval. For serving, set labels to None. default_inputs: a dict of string to Tensor or SparseTensor, specifying the input placeholders (if any) that this input_fn expects to be fed. Typically, this is used by a serving input_fn, which expects to be fed serialized tf.Example protos.