reason for the rmse via data science on gcp
we discussed the need for a metric that is independent of threshold and captures the full spectrum of probabilities. The Area Under the Curve (AUC) measure that is reported by TensorFlow does this, but it is not as sensitive to model improvements as the Root Mean Squared Error (RMSE) that we started with in Chapter 7. For comparison purposes, therefore, it would be good to also compute the RMSE.